Add new `squiz::is_always_nothrow_connectable_v<Sender, Env...>` query.
This allows senders to declare that their connect() methods will be
noexcept regardless of the receiver type that is passed to them.
This allows operations that need to call connect() during their
execution (i.e. after start() has been called) to correctly compute
the completion-signatures based on whether or not connect() might
throw.