Add variant_child_operation method overloads that take a runtime index
Added overloads to `start()`, `request_stop()` and `destruct()` that
takes a runtime index value and dispatches to the call to the
corresponding child operation state.

Added a new `detail::dispatch_index()` helper function for invoking an
instantiation of a lambda with the std::integral_constant type
corresponding to the runtime value of an index. This is used in the
implementation of the new variant_child_operation methods above.