Add the let_value sender algorithm.
This implements a lock-free cancellable version that forwards
stop-requests on to the active child operation.

Note that even if source sender and successor-senders all are not able to complete
with set_stopped() the resulting completion-signatures must include
set_stopped() as the let_value() algorithm can itself complete with set_stopped().