Add 'stop_token_sender' adaptor
This adaptor turns a stop-token into a sender which completes with
value_t<> if a stop-request is received, but can also be cancelled
by a call to request_stop() on the operation-state.

This can be used in conjunction with stop_when() to allow an operation
to be cancelled via a stop-token.

Added some other utilities to support this:
- stoppable_token concept
- unstoppable_token concept
- detail::stop_callback_type