Disable optimization in stop_when()
The optimization in stop_when() that skips starting the trigger
operation if the source operation completes synchronously was causing
some issues in some cases when used with counting_scope as if the
trigger operation was a nest() operation then skipping starting the
trigger operation means that the ref-count is not released as early
and so can lead to some deadlock situations.

Need to investigate more issues around when operation-states should
be destroyed to ensure that they are destroyed as early as possible.