Add `detail::bitset` class.
This type provides a subset of the functionality of the `std::bitset`
type but, importantly, provides the countr_zero() function for
scanning the bitset for the index of the first set bit in the bitset.

This is useful for a hashed priority queue data-structure I've been
working on to efficiently find which bins are non-empty.