Improvements to binned_time_priority_queue
Added documentation of the data-structure and methods and describe the
intended usage and expected computational complexity.

Tweaked the concepts to require that the \c GetTimestampFunc was
\c std::regular_invocable - i.e. that invoking it multiple times
gives the same result.

Added an implementation for the move-constructor as the default
implementation of the move-constructor would have left the source
object in an inconsistent state.

Added an \c empty() method for checking if the queue contains any
items or not.