Add atomic_intrusive_queue data-structure.
This allows lock-free enqueuing of items from multiple threads and a
single thread to dequeue all items from the queue.
Also allows setting the queue to the 'inactive' state to signal to
subsequent enqueuers that they need to do some work after enqueueing
an item to wake up the consumer of the items so that it will dequeue
them.

Also added `intrusive_queue` and `intrusive_stack` types that provide
intrusive singly-linked-list data-structures.
The `intrusive_queue` data-structure allows pushing to both the front
and back, and popping from the front.
The `intrusive_stack` data-structure only allows pushing/popping from
the front of the list.

Data structures based on code from libunifex.
https://github.com/facebookexperimental/libunifex/