Add the `squiz::linuxos::iouring_context` class.
This type is a wrapper around an io_uring data-structure for
issuing asynchronous syscalls to the OS.

It currently only implements the time_scheduler interface methods:
schedule(), schedule_at(), schedule_after(), now(), but provides the
necessary infrastructure for issuing async syscalls.

The schedule() operation supports cancellation, although a
stop-request is ignored if it is not made on the I/O thread.

The schedule_at/after() operations also support cancellation.
A stop-request from a remote thread will remotely enqueue a task
to be run on the I/O thread to remove the timer from the priority
queue.

The iouring_operation base-class handles generic cancellation of
submitted SQE operations, including handling either a stop_request
coming from the I/O thread or another thread.

Implemented the openat() and close() operations on the
iouring_context::scheduler and add a test that shows they can be
executed with and without cancellation.

These are mostly just to show a functional code-path for submitting
SQE-based operations and processing completions and handling
cancellation. Ideally they would be refactored to use strong types
for accessing different kinds of resources instead.

Note that some io_uring optimizations are disabled by default. In
particular the optimisations that put constraints on calling run()
on the same thread tht constructed the io_uring object, which would
be a bit restrictive for some test-use cases which want to construct
the context in main, but then drive it from a background thread.